+++ src/SexyAppFramework/SoundManager.h
@@ -29,7 +29,7 @@ class SoundManager
 
 	virtual void SetVolume(double theVolume) = 0;
 	virtual bool SetBaseVolume(unsigned int theSfxID, double theBaseVolume) = 0;
-	virtual bool SetBasePan(unsigned int theSfxID, int theBasePan) = 0;
+	virtual bool SetBasePan(unsigned int theSfxID, float theBasePan) = 0;
 
 	virtual SoundInstance *GetSoundInstance(unsigned int theSfxID) = 0;
